fix: distinguish untracked sink from sink with no tags.
Given a sink, the tag analysis returns whether the analysis kept track
of all possible sources that can lead to the sink or not. If not, then
the worst case should be assumed (i.e., all sources might reach the
sink). When, the analysis knows about a sink it is still possible that
the set of possible sources is empty. The meaning of this is that the
analysis cannot connect a source with a sink. This is typically a case
to inspect manually. Possible reasons are: a source is not recognized
(see sources-sinks.config for all supported sources), or the source is
defined in another library which is not available to the Tag analysis.